Improve random definitions given for bottom cards
Why - the card definition would overlap because before it checked if
the card had a string of "term" And so it would only work if the card
had term. First run it would work because all cards are set to term but
second run would glitch because cards are no longer term.